Move class for IT to separate file
