+++ tests/system/aiplatform/test_e2e_vertex_model.py
@@ -22,76 +22,9 @@ import pandas as pd
 import torch
 
 from google.cloud import aiplatform
-from google.cloud.aiplatform.experimental.vertex_model import base
 
 from tests.system.aiplatform import e2e_base
-
-
-class LinearRegression(base.VertexModel, torch.nn.Module):
-    def __init__(self, input_size: int, output_size: int):
-        base.VertexModel.__init__(self, input_size=input_size, output_size=output_size)
-        torch.nn.Module.__init__(self)
-        self.linear = torch.nn.Linear(input_size, output_size)
-
-    def forward(self, x):
-        return self.linear(x)
-
-    def train_loop(self, dataloader, loss_fn, optimizer):
-        for batch, (X, y) in enumerate(dataloader):
-            pred = self.predict(X.float())
-            loss = loss_fn(pred.float(), y.float())
-
-            optimizer.zero_grad()
-            loss.backward()
-            optimizer.step()
-
-    def fit(
-        self, data: pd.DataFrame, target_column: str, epochs: int, learning_rate: float
-    ):
-        feature_columns = list(data.columns)
-        feature_columns.remove(target_column)
-
-        features = torch.tensor(data[feature_columns].values)
-        target = torch.tensor(data[target_column].values)
-
-        dataloader = torch.utils.data.DataLoader(
-            torch.utils.data.TensorDataset(features, target),
-            batch_size=10,
-            shuffle=True,
-        )
-
-        loss_fn = torch.nn.MSELoss()
-        optimizer = torch.optim.SGD(self.parameters(), lr=learning_rate)
-
-        for t in range(epochs):
-            self.train_loop(dataloader, loss_fn, optimizer)
-
-    def predict(self, data):
-        return self.forward(data)
-
-    # Implementation of predict_payload_to_predict_input(), which converts a predict_payload object to predict() inputs
-    def predict_payload_to_predict_input(self, instances):
-        feature_columns = ["feat_1", "feat_2"]
-        data = pd.DataFrame(instances, columns=feature_columns)
-        torch_tensor = torch.tensor(data[feature_columns].values).type(
-            torch.FloatTensor
-        )
-        return torch_tensor
-
-    # Implementation of predict_input_to_predict_payload(), which converts predict() inputs to a predict_payload object
-    def predict_input_to_predict_payload(self, parameter):
-        return parameter.tolist()
-
-    # Implementation of predict_output_to_predict_payload(), which converts the predict() output to a predict_payload object
-    def predict_output_to_predict_payload(self, output):
-        return output.tolist()
-
-    # Implementation of predict_payload_to_predict_output, which takes a predict_payload object containing predictions and
-    # converts it to the type of output expected by the user-written class.
-    def predict_payload_to_predict_output(self, predictions):
-        data = pd.DataFrame(predictions)
-        torch_tensor = torch.tensor(data.values).type(torch.FloatTensor)
-        return torch_tensor
+from tests.system.aiplatform import test_vertex_model
 
 
 @pytest.mark.usefixtures("prepare_staging_bucket", "delete_staging_bucket", "teardown")
@@ -120,7 +53,7 @@ class TestEndToEndVertexModel(e2e_base.TestEndToEnd):
         )
 
         # Train using VertexModel interface
-        my_model = LinearRegression(2, 1)
+        my_model = test_vertex_model.LinearRegression(2, 1)
         my_model.remote = training_remote
 
         my_model.fit(df, "target", 1, 0.1)
